+++ argo-plugin/src/main/kotlin/io/github/leofuso/argo/plugin/columba/SpecificCompilerRunnerApplication.kt
@@ -1,0 +1,3 @@
+package io.github.leofuso.argo.plugin.columba
+
+class SpecificCompilerRunnerApplication
