+++ argo-plugin/src/main/kotlin/io/github/leofuso/argo/plugin/columba/exception/NonDeterministicSchemaResolutionException.kt
@@ -1,0 +1,6 @@
+package io.github.leofuso.argo.plugin.columba.exception
+
+class NonDeterministicSchemaResolutionException : RuntimeException {
+    constructor(message: String) : super(message)
+    constructor(message: String, exception: Exception) : super(message, exception)
+}
