--- argo-plugin/src/main/kotlin/io/github/leofuso/argo/plugin/compiler/ClasspathClassLoaderFactory.kt
@@ -1,21 +1,0 @@
-package io.github.leofuso.argo.plugin.compiler
-
-import org.gradle.internal.classloader.VisitableURLClassLoader
-import java.io.File
-import java.net.URLClassLoader
-
-inline fun <reified T> T.urlClassLoader(files: Set<File>): URLClassLoader = files.map { it.toURI().toURL() }
-    .toTypedArray()
-    .let {
-        URLClassLoader.newInstance(it, T::class.java.classLoader)
-    }
-
-fun ClassLoader.loadURLs(files: Set<File>): ClassLoader = files.mapNotNull {
-    val uri = it.toURI()
-    uri.toURL()
-}.fold(this) { loader, url ->
-    if (loader is VisitableURLClassLoader) {
-        loader.addURL(url)
-    }
-    return this
-}
