--- argo-plugin/src/main/kotlin/io/github/leofuso/argo/plugin/compiler/SpecificCompilerRunnerApplication.kt
@@ -1,3 +1,0 @@
-package io.github.leofuso.argo.plugin.compiler
-
-class SpecificCompilerRunnerApplication
