--- argo-plugin/src/main/kotlin/io/github/leofuso/argo/plugin/compiler/arguments/Argument.kt
@@ -1,7 +1,0 @@
-package io.github.leofuso.argo.plugin.compiler.arguments
-
-internal sealed interface Argument {
-
-    fun args(): List<String>
-
-}
