--- argo-plugin/src/main/kotlin/io/github/leofuso/argo/plugin/compiler/exception/NonDeterministicSchemaResolutionException.kt
@@ -1,6 +1,0 @@
-package io.github.leofuso.argo.plugin.compiler.exception
-
-class NonDeterministicSchemaResolutionException : RuntimeException {
-    constructor(message: String) : super(message)
-    constructor(message: String, exception: Exception) : super(message, exception)
-}
