+++ argo-plugin/src/test/kotlin/io/github/leofuso/argo/plugin/parser/DependencyResolutionParserTest.kt
@@ -2,8 +2,8 @@
 
 package io.github.leofuso.argo.plugin.parser
 
-import io.github.leofuso.argo.plugin.compiler.parser.DefaultSchemaParser
-import io.github.leofuso.argo.plugin.compiler.parser.DependencyGraphAwareSchemaParser
+import io.github.leofuso.argo.plugin.columba.parser.DefaultSchemaParser
+import io.github.leofuso.argo.plugin.columba.parser.DependencyGraphAwareSchemaParser
 import io.github.leofuso.argo.plugin.fixtures.FileTreeParameterResolver
 import io.github.leofuso.argo.plugin.fixtures.annotation.SchemaParameter
 import io.github.leofuso.argo.plugin.fixtures.loadResource
