+++ argo-plugin/src/test/kotlin/io/github/leofuso/argo/plugin/parser/ProtocolParserTest.kt
@@ -1,7 +1,7 @@
 package io.github.leofuso.argo.plugin.parser
 
-import io.github.leofuso.argo.plugin.compiler.parser.DefaultSchemaParser
-import io.github.leofuso.argo.plugin.compiler.parser.DependencyGraphAwareSchemaParser
+import io.github.leofuso.argo.plugin.columba.parser.DefaultSchemaParser
+import io.github.leofuso.argo.plugin.columba.parser.DependencyGraphAwareSchemaParser
 import io.github.leofuso.argo.plugin.fixtures.FileTreeParameterResolver
 import io.github.leofuso.argo.plugin.fixtures.loadResource
 import org.apache.avro.Protocol
