+++ columba-cli/src/main/kotlin/io/github/leofuso/columba/cli/exceptions/NonDeterministicSchemaResolutionException.kt
@@ -1,0 +1,6 @@
+package io.github.leofuso.columba.cli.exceptions
+
+class NonDeterministicSchemaResolutionException : RuntimeException {
+    constructor(message: String) : super(message)
+    constructor(message: String, exception: Exception) : super(message, exception)
+}
