--- columba-cli/src/main/kotlin/io/github/leofuso/columba/compiler/cli/Constants.kt
@@ -1,33 +1,0 @@
-package io.github.leofuso.columba.compiler.cli
-
-/**
- * Extensions
- */
-const val CLASS_EXTENSION = "class"
-const val SCHEMA_EXTENSION = "avsc"
-const val PROTOCOL_EXTENSION = "avpr"
-const val VELOCITY_TEMPLATE_EXTENSION = "vm"
-const val IDL_EXTENSION = "avdl"
-const val JAVA_EXTENSION = "java"
-const val AVRO_EXTENSION_NAME = "avro"
-
-/**
- * Compreesed extensions
- */
-const val JAR_EXTENSION = "jar"
-const val ZIP_EXTENSION = "zip"
-
-/**
- * The namespace separator.
- */
-const val NAMESPACE_SEPARATOR = "."
-
-/**
- * The extension separator.
- */
-const val EXTENSION_SEPARATOR = "."
-
-/**
- * The Unix separator.
- */
-const val UNIX_SEPARATOR = "/"
