--- columba-cli/src/main/kotlin/io/github/leofuso/columba/compiler/cli/exceptions/NonDeterministicSchemaResolutionException.kt
@@ -1,6 +1,0 @@
-package io.github.leofuso.columba.compiler.cli.exceptions
-
-class NonDeterministicSchemaResolutionException : RuntimeException {
-    constructor(message: String) : super(message)
-    constructor(message: String, exception: Exception) : super(message, exception)
-}
